add more code

+++ chapter-06/listing_03_rust_psuedocode_NO_COMPILE/src/lib.rs
@@ -1,4 +1,4 @@
-pub fn sum(line: &str) -> i64 {
+pub fn sum(line: &str) -> i32 {
   let data = parse_as_json(line);
 
   data.value + data.name.len()
